requirements.txt
Pillow
pickle
opencv-python
pandas
sklearn
getopt
numpy
imageio
imblearn
gdal